refactor(taiko-client): split pkg/utils by concern and move LoadEnv to testutils
